+++ app/src/main/java/com/barbuceanuconstantin/proiectlicenta/MainActivity.kt
@@ -6,11 +6,7 @@ import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.annotation.RequiresApi
 import com.barbuceanuconstantin.proiectlicenta.ui.theme.ProiectLicentaTheme
-import com.barbuceanuconstantin.proiectlicenta.view.screen.showTransactionDialog
-import com.barbuceanuconstantin.proiectlicenta.view.screen.transactionsComposableScreen
-import com.barbuceanuconstantin.proiectlicenta.view.screenmodules.screen1
-import com.barbuceanuconstantin.proiectlicenta.view.screenmodules.screen2
-import com.barbuceanuconstantin.proiectlicenta.view.screenmodules.showMenu
+import com.barbuceanuconstantin.proiectlicenta.view.screen.screen1
 
 //Branch ecrane-individuale//
