+++ app/src/main/java/com/barbuceanuconstantin/proiectlicenta/view/screen/MainScreens.kt
@@ -1,0 +1,71 @@
+package com.barbuceanuconstantin.proiectlicenta.view.screen
+
+import android.os.Build
+import androidx.annotation.RequiresApi
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.mutableIntStateOf
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.snapshots.SnapshotStateList
+import androidx.compose.runtime.toMutableStateList
+import com.barbuceanuconstantin.proiectlicenta.data.model.Subcategory
+import com.barbuceanuconstantin.proiectlicenta.data.model.Tranzactie
+import com.barbuceanuconstantin.proiectlicenta.lTrA
+import com.barbuceanuconstantin.proiectlicenta.lTrD
+import com.barbuceanuconstantin.proiectlicenta.lTrP
+import com.barbuceanuconstantin.proiectlicenta.subcategorysPredefiniteActive
+import com.barbuceanuconstantin.proiectlicenta.subcategorysPredefiniteDatorii
+import com.barbuceanuconstantin.proiectlicenta.subcategorysPredefinitePasive
+
+private var lTranzactiiActive: SnapshotStateList<Tranzactie> = lTrA
+private var lTranzactiiPasive: SnapshotStateList<Tranzactie> = lTrP
+private var lTranzactiiDatorii: SnapshotStateList<Tranzactie> = lTrD
+
+private var listSubcategoriesRevenue = subcategorysPredefiniteActive.map {
+    Subcategory(name = it.key.toString(), items = it.value.toMutableStateList())
+}.toMutableStateList()
+private var listSubcategoriesExpenses = subcategorysPredefinitePasive.map {
+    Subcategory(name = it.key.toString(), items = it.value.toMutableStateList())
+}.toMutableStateList()
+private var listSubcategoriesDebts = subcategorysPredefiniteDatorii.map {
+    Subcategory(name = it.key.toString(), items = it.value.toMutableStateList())
+}.toMutableStateList()
+@Composable
+fun screen0() {
+    var sumRevenue: Float = 0f;
+    var sumExpenses: Float = 0f;
+    var sumDebt: Float = 0f;
+    principalComposableScreen(sumRevenue, sumExpenses, sumDebt)
+}
+
+@RequiresApi(Build.VERSION_CODES.O)
+@Composable
+fun screen1() {
+    val showA = mutableStateOf(true)
+    val showP = mutableStateOf(true)
+    val showD = mutableStateOf(true)
+    val addButton = mutableStateOf(false)
+    val index = mutableIntStateOf(-1)
+    val sem = mutableIntStateOf(-1)
+    val updateTransactionButton = mutableStateOf(false)
+
+    transactionsComposableScreen(showA, showP, showD, addButton, lTranzactiiActive, lTranzactiiPasive, lTranzactiiDatorii, index, sem, updateTransactionButton)
+}
+@Composable
+fun screen2() {
+    val showA = mutableStateOf(true)
+    val showP = mutableStateOf(true)
+    val showD = mutableStateOf(true)
+    val addButton = mutableStateOf(false)
+
+    categoriesComposableScreen(showA, showP, showD, addButton, listSubcategoriesRevenue, listSubcategoriesExpenses, listSubcategoriesDebts)
+}
+@Composable
+fun screen4() {
+    val fab = mutableStateOf(false)
+
+    fixedBudgetsComposableScreen(fab)
+}
+@Composable
+fun screen7() {
+    calendarComposableScreen()
+}
