+++ app/src/main/java/com/barbuceanuconstantin/proiectlicenta/view/screenmodules/MenuScreensSwipeableTabRows.kt
@@ -37,30 +37,19 @@ import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableIntStateOf
-import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
-import androidx.compose.runtime.snapshots.SnapshotStateList
-import androidx.compose.runtime.toMutableStateList
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.vector.ImageVector
 import androidx.compose.ui.platform.LocalConfiguration
 import androidx.compose.ui.res.stringResource
 import com.barbuceanuconstantin.proiectlicenta.R
-import com.barbuceanuconstantin.proiectlicenta.data.model.Subcategory
-import com.barbuceanuconstantin.proiectlicenta.data.model.Tranzactie
-import com.barbuceanuconstantin.proiectlicenta.lTrA
-import com.barbuceanuconstantin.proiectlicenta.lTrD
-import com.barbuceanuconstantin.proiectlicenta.lTrP
-import com.barbuceanuconstantin.proiectlicenta.subcategorysPredefiniteActive
-import com.barbuceanuconstantin.proiectlicenta.subcategorysPredefiniteDatorii
-import com.barbuceanuconstantin.proiectlicenta.subcategorysPredefinitePasive
-import com.barbuceanuconstantin.proiectlicenta.view.screen.calendarComposableScreen
-import com.barbuceanuconstantin.proiectlicenta.view.screen.categoriesComposableScreen
-import com.barbuceanuconstantin.proiectlicenta.view.screen.fixedBudgetsComposableScreen
-import com.barbuceanuconstantin.proiectlicenta.view.screen.principalComposableScreen
-import com.barbuceanuconstantin.proiectlicenta.view.screen.transactionsComposableScreen
+import com.barbuceanuconstantin.proiectlicenta.view.screen.screen0
+import com.barbuceanuconstantin.proiectlicenta.view.screen.screen1
+import com.barbuceanuconstantin.proiectlicenta.view.screen.screen2
+import com.barbuceanuconstantin.proiectlicenta.view.screen.screen4
+import com.barbuceanuconstantin.proiectlicenta.view.screen.screen7
 
 data class TabItem(
     val title: String,
@@ -68,61 +57,7 @@ data class TabItem(
     val selectedIcon: ImageVector
 )
 
-private var lTranzactiiActive: SnapshotStateList<Tranzactie> = lTrA
-private var lTranzactiiPasive: SnapshotStateList<Tranzactie> = lTrP
-private var lTranzactiiDatorii: SnapshotStateList<Tranzactie> = lTrD
-
-private var listSubcategoriesRevenue = subcategorysPredefiniteActive.map {
-    Subcategory(name = it.key.toString(), items = it.value.toMutableStateList())
-}.toMutableStateList()
-private var listSubcategoriesExpenses = subcategorysPredefinitePasive.map {
-    Subcategory(name = it.key.toString(), items = it.value.toMutableStateList())
-}.toMutableStateList()
-private var listSubcategoriesDebts = subcategorysPredefiniteDatorii.map {
-    Subcategory(name = it.key.toString(), items = it.value.toMutableStateList())
-}.toMutableStateList()
-
-@Composable
-fun screen0() {
-    var sumRevenue: Float = 0f;
-    var sumExpenses: Float = 0f;
-    var sumDebt: Float = 0f;
-    principalComposableScreen(sumRevenue, sumExpenses, sumDebt)
-}
-
 @RequiresApi(Build.VERSION_CODES.O)
-@Composable
-fun screen1() {
-    var showA = mutableStateOf(true)
-    var showP = mutableStateOf(true)
-    var showD = mutableStateOf(true)
-    var addButton = mutableStateOf(false)
-    val index = mutableIntStateOf(-1)
-    val sem = mutableIntStateOf(-1)
-    val updateTransactionButton = mutableStateOf(false)
-
-    transactionsComposableScreen(showA, showP, showD, addButton, lTranzactiiActive, lTranzactiiPasive, lTranzactiiDatorii, index, sem, updateTransactionButton)
-}
-@Composable
-fun screen2() {
-    var showA = mutableStateOf(true)
-    var showP = mutableStateOf(true)
-    var showD = mutableStateOf(true)
-    var addButton = mutableStateOf(false)
-
-    categoriesComposableScreen(showA, showP, showD, addButton, listSubcategoriesRevenue, listSubcategoriesExpenses, listSubcategoriesDebts)
-}
-@Composable
-fun screen4() {
-    var fab = mutableStateOf(false)
-
-    fixedBudgetsComposableScreen(fab)
-}
-@Composable
-fun screen7() {
-    calendarComposableScreen()
-}
-
 @Composable
 fun showMenu() {
     val tabItems = listOf(
